requirements.txt
ultralytics
opencv-python
opencv-contrib-python
boto3
pillow
shapely
gdown
glob2
python-dotenv
tensorboard
matplotlib
tqdm
psutil
scipy
bs4
google
protobuf
absl-py
botocore
s3transfer